[refactor] 폴더 및 클래스 네이밍
