+++ src/main/java/com/jinddung2/givemeticon/domain/oauth/domain/AuthTokenGenerator.java
@@ -1,6 +1,6 @@
 package com.jinddung2.givemeticon.domain.oauth.domain;
 
-import com.jinddung2.givemeticon.common.security.provider.JwtTokenProvider;
+import com.jinddung2.givemeticon.common.security.utils.JwtTokenUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -13,7 +13,7 @@ public class AuthTokenGenerator {
     private static final long ACCESS_TOKEN_EXPIRE_TIME = 1000 * 60 * 30;            // 30분
     private static final long REFRESH_TOKEN_EXPIRE_TIME = 1000 * 60 * 60 * 24 * 7;  // 7일
 
-    private final JwtTokenProvider jwtTokenProvider;
+    private final JwtTokenUtil jwtTokenUtil;
 
     public AuthToken generate(Integer userId) {
         long now = (new Date()).getTime();
@@ -21,13 +21,13 @@ public class AuthTokenGenerator {
         Date refreshTokenExpiredDate = new Date(now + REFRESH_TOKEN_EXPIRE_TIME);
 
         String subject = userId.toString();
-        String accessToken = jwtTokenProvider.generate(subject, accessTokenExpiredDate);
-        String refreshToken = jwtTokenProvider.generate(subject, refreshTokenExpiredDate);
+        String accessToken = jwtTokenUtil.generate(subject, accessTokenExpiredDate);
+        String refreshToken = jwtTokenUtil.generate(subject, refreshTokenExpiredDate);
 
         return AuthToken.of(accessToken, refreshToken, BEARER_TYPE, ACCESS_TOKEN_EXPIRE_TIME / 1000L);
     }
 
     public Long extractUserId(String accessToken) {
-        return Long.valueOf(jwtTokenProvider.extractSubject(accessToken));
+        return Long.valueOf(jwtTokenUtil.extractSubject(accessToken));
     }
 }
